Correcciones varias

+++ VentaDeuda/src/main/java/com/fiuba/VentaDeuda/domain/Deuda.java
@@ -5,11 +5,17 @@ import javax.validation.constraints.NotNull;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fiuba.VentaDeuda.enums.EstadoDeuda;
 import lombok.Data;
+import org.springframework.cglib.core.Local;
+import org.springframework.cglib.core.internal.LoadingCache;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.*;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.math.BigInteger;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.time.LocalDate;
 import java.util.Date;
 
 @Entity
@@ -21,8 +27,9 @@ public class Deuda implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
+    //El valor es el monto total de la deuda a cobrar. Es la cantidad que debe cobrar el usuario que compra la deuda.
     @NotNull
-    private BigInteger valor;
+    private BigDecimal valor;
 
     @ManyToOne
     @JoinColumn
@@ -32,8 +39,9 @@ public class Deuda implements Serializable {
     @JoinColumn
     private Usuario vendedor;
 
+    //El precio es el monto por el cual la deuda es vendida. Representa una fraccion del valor.
     @NotNull
-    private BigInteger precio;
+    private BigDecimal precio;
 
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
@@ -53,4 +61,5 @@ public class Deuda implements Serializable {
         this.setVendedor(usuario);
         this.setEstado(EstadoDeuda.NO_VENDIDO);
     }
+
 }
