+++ VentaDeuda/src/main/java/com/fiuba/VentaDeuda/dto/usuario/SaldoUsuarioRequest.java
@@ -4,6 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
@@ -11,6 +14,6 @@ public class SaldoUsuarioRequest {
 
     private String userName;
 
-    private int saldo;
+    private BigDecimal saldo;
 
 }
